+++ lib/runtime.sh
@@ -5,6 +5,7 @@ install_runtime_files() {
   create_service_user adgvpn /var/lib/adguardvpn
   create_root_dir /var/log/myvpn 0755
   create_root_dir /var/lib/vpn-rotate 0700
+  create_root_dir /var/lib/watchdogvpn 0755
 
   create_config_if_missing "$ROOT_DIR/examples/adguardvpn.env.example" /etc/adguardvpn.env 0644
   create_config_if_missing "$ROOT_DIR/examples/vpn-domain-bypass.conf.example" /etc/vpn-domain-bypass.conf 0644
@@ -14,6 +15,7 @@ install_runtime_files() {
   install_root_file "$ROOT_DIR/bin/vpn_auth_check" /usr/local/bin/vpn_auth_check 0755
   install_root_file "$ROOT_DIR/bin/vpn_dns_rescue" /usr/local/bin/vpn_dns_rescue 0755
   install_root_file "$ROOT_DIR/bin/vpn_dnsctl" /usr/local/bin/vpn_dnsctl 0755
+  install_root_file "$ROOT_DIR/bin/vpn_manual_state" /usr/local/bin/vpn_manual_state 0755
   install_root_file "$ROOT_DIR/bin/vpn_notify" /usr/local/bin/vpn_notify 0755
   install_root_file "$ROOT_DIR/bin/vpn_truth_check" /usr/local/bin/vpn_truth_check 0755
   install_root_file "$ROOT_DIR/bin/vpnctl" /usr/local/bin/vpnctl 0755